+++ inbm/dispatcher-agent/tests/unit/test_provision_target.py
@@ -11,6 +11,10 @@ TEST_XML = '<manifest><type>cmd</type><cmd>provisionNode</cmd><provisionNode>' \
            '<fetch>https://www.repo.com/provision.tar</fetch><signature>signature</signature>' \
            '</provisionNode></manifest>'
 
+TEST_XML_INVAID_SIG_VERSION = '<manifest><type>cmd</type><cmd>provisionNode</cmd><provisionNode>' \
+                              '<fetch>https://www.repo.com/provision.tar</fetch><signature>signature</signature>' \
+                              '<signature_version>384abc</signature_version></provisionNode></manifest>'
+
 BAD_XML = '<manifest><type>cmd</type><cmd>provisionNode</cmd><provisionNode>' \
     '<path>https://www.repo.com/provision.tar</path><signature>signature</signature>' \
     '</provisionNode></manifest>'
@@ -34,6 +38,7 @@ class TestProvisionTarget(TestCase):
     def setUp(self, mock_dispatcher):
         self.mocked_dispatcher = mock_dispatcher
         self.parsed = XmlHandler(xml=TEST_XML, is_file=False, schema_location=TEST_SCHEMA_LOCATION)
+        self.parsed_sig_version = XmlHandler(xml=TEST_XML_INVAID_SIG_VERSION, is_file=False, schema_location=TEST_SCHEMA_LOCATION)
 
     def test_successfully_verify_files(self):
         files = [blob_files, cert_files]
@@ -65,6 +70,12 @@ class TestProvisionTarget(TestCase):
     def test_successfully_install(self, mock_download, mock_extract):
         p = ProvisionTarget(TEST_XML, self.mocked_dispatcher, TEST_SCHEMA_LOCATION)
         p.install(self.parsed)
+
+    @patch('dispatcher.provision_target.extract_files_from_tar', return_value=([blob_files, cert_files], test_tar))
+    @patch('dispatcher.provision_target.download')
+    def test_successfully_install_with_invalid_signature_version(self, mock_download, mock_extract):
+        p = ProvisionTarget(TEST_XML_INVAID_SIG_VERSION, self.mocked_dispatcher, TEST_SCHEMA_LOCATION)
+        p.install(self.parsed_sig_version)
 
     @patch('dispatcher.provision_target.extract_files_from_tar', return_value=(None, 'test.tar'))
     @patch('dispatcher.provision_target.download')