+++ inbm-vision/Changelog.md
@@ -3,11 +3,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## 3.0.8 - 2022-01-24
+
+### Fixed
+- 504494, HSD 15010332252 - Fix provision file not removed in /opt/xlink_provision when failed to reset device due to no matching device id
+- 506035, HSD 15010309026 - Update node apparmor policy by adding /sys/kernel/xlink_secure/debug
+- HSD 15010516458 - Fix storage checks failed during mender file transferring in SOTA/POTA
+- 506874 - Fix node config schema error when upgrading from v2.16.0 to v2.17.2
+- 507151 - Add lower bound, upper bound, and default values to node-agent configuration values.
+- 507593 - Add lower bound, upper bound, and default values to vision-agent configuration values.
+
 ## 3.0.7 - 2022-01-04
 
 ## Fixed
  - Build from normal Internet connection fixed.
-
 
 ## 3.0.6 - 2021-12-31
 This is the first open source release.